Finishing  to clean the folders
