+++ FinderLost/ContainerDI/PreviewProvider+StoreMock.swift
@@ -1,12 +1,13 @@
 //
-//  StoreMock.swift
+//  ContainerDI+Store.swift
 //  FinderLost
 //
-//  Created by Andres Felipe Alzate Restrepo on 28/1/23.
+//  Created by Andres Felipe Alzate Restrepo on 11/2/23.
 //
 
 import Redux
 import SwiftUI
+import Factory
 
 extension PreviewProvider {
     static var storeMock: Store<FinderLost> {
