+++ Packages/Data/Sources/Data/ContainerDI/ContainerDI+Firebase.swift
@@ -1,0 +1,17 @@
+//
+//  ContainerDI+Firebase.swift
+//  
+//
+//  Created by Andres Felipe Alzate Restrepo on 11/2/23.
+//
+
+import Factory
+
+import Firebase
+import GoogleSignIn
+
+extension Container {
+    static let firebaseApp = Factory<FirebaseApp?>(scope: .singleton) { FirebaseApp.app() }
+    static let firebaseAuth = Factory<Auth>(scope: .singleton) { Auth.auth() }
+    static let googleSignIn = Factory<GIDSignIn>(scope: .singleton) { GIDSignIn.sharedInstance }
+}
