--- Packages/Data/Sources/Data/Data.swift
@@ -1,6 +1,0 @@
-public struct Data {
-    public private(set) var text = "Hello, World!"
-
-    public init() {
-    }
-}
